phishnet-backend/requirements.txt
fastapi
uvicorn
requests
openai
torch
transformers
sentence-transformers
pydantic
beautifulsoup4
python-dotenv
scikit-learn
